vite, moca effect, calc size move to rust, web player

+++ src/js/listItems.js
@@ -1,6 +1,9 @@
+import { ui } from "../main.js";
 import { LIST_ITEM_TYPES } from "./enums/listItems.js";
+import { isAudio } from "./functions/fileFormats.js";
 import { fromHtml } from "./functions/html.js";
 import { deletePath } from "./functions/listFuncs.js";
+
 
 
 
@@ -31,16 +34,37 @@ const folderElement = (item) => {
     return li;
 }
 const fileElement = (item) => {
-    const { name, size, modified, readonly, hidden, path } = item;
+    const { name, size, modified, readonly, hidden, path, fileType } = item;
     const li = document.createElement("li");
     li.dataset.type = LIST_ITEM_TYPES.FILE;
     li.classList.add("folder__list-item");
+    li.dataset.name = name;
+    li.dataset.path = path;
+    li.dataset.size = size;
     li.innerHTML = fileHtml(item);
-    li.querySelector(".delete-button")?.addEventListener("click", async (e) => {
-        e.stopPropagation();
-        await deletePath(path);
-        li.remove();
-    })
+    li.addEventListener("click", async (event) => {
+        if (event.target.closest(".delete-button")) {
+            event.stopPropagation();
+            await deletePath(path);
+            li.remove();
+            console.log("delete", path)
+        }
+        // else {
+        //     if (isAudio(item)) {
+
+        //         ui.startPlayer(item);
+        //     }
+        // }
+    });
+    li.addEventListener("dblclick", async (event) => {
+        console.log(isAudio(item), item)
+
+        if (isAudio(item)) {
+
+            ui.startPlayer(item);
+        }
+
+    });
     return li;
 }
 export const driveElement = (drive) => {
